% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomizer.R
\name{atomizer}
\alias{atomizer}
\title{atomizer}
\usage{
atomizer(path_ref, path_new, funcs)
}
\arguments{
\item{path_ref}{path to reference package, end of path is pkg name}

\item{path_new}{path to new package, end of path is pkg name}

\item{funcs}{function(s) to extract}
}
\description{
atomizer
}
\examples{
\dontrun{
atomizer(path_ref = "../rredlist", path_new = "../foobar", funcs = "rl_citation")
devtools::install("../foobar")
library("foobar")
foobar::rl_citation()
}
}

