% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funcs.R
\name{extract_funcs}
\alias{extract_funcs}
\title{Get all functions needed for a given function}
\usage{
extract_funcs(path, x)
}
\arguments{
\item{path}{Path to a package. Default: \code{.}. required}

\item{x}{One or more functions. required.}
}
\value{
A list
}
\description{
Get all functions needed for a given function
}
\examples{
\dontrun{
res <- scattalyze("../rredlist")
(funcs <- find_deps(res, fxns = "rl_citation"))
extract_funcs(path="../rredlist", funcs)
}
}

