% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imports.R
\name{get_imports}
\alias{get_imports}
\title{Get imported packages and their functions}
\usage{
get_imports(path, x)
}
\arguments{
\item{path}{Path to a package. Default: \code{.}. required}

\item{x}{One or more functions. required.}
}
\value{
A list
}
\description{
Get imported packages and their functions
}
\examples{
\dontrun{
res <- scattalyze("../rredlist")
(funcs <- find_deps(res, fxns = "rl_citation"))
extract_funcs(path="../rredlist", funcs)
get_imports(path="../rredlist", funcs)
}
}
