% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattalyze.R
\name{scattalyze}
\alias{scattalyze}
\title{Scan package and get functions, and their dependencies, etc}
\usage{
scattalyze(path = ".", include_base = FALSE)
}
\arguments{
\item{path}{Path to a package. Default: \code{.}}

\item{include_base}{Whether to include functions from the 'base' package.
Default: \code{FALSE}}
}
\value{
A function_map S3 object (a list basically)
}
\description{
Scan package and get functions, and their dependencies, etc
}
\examples{
\dontrun{
(res <- scattalyze("../rredlist"))
# data
res$data
# exports
res$exports
}
}

